모달창 컴포넌트,  모달창내 방만들기 컴포넌트 합침.
