+++ coco/db.json
@@ -29,6 +29,16 @@
       "roomName": "ㅗㅗ",
       "category": "초기값^^",
       "id": 6
+    },
+    {
+      "roomName": "ㄴㄴㅇ",
+      "category": "카테2",
+      "id": 7
+    },
+    {
+      "roomName": "",
+      "category": "초기값^^",
+      "id": 8
     }
   ],
   "comments": [
