+++ coco/src/components/Main/ModalCreateRoom.jsx
@@ -1,10 +1,35 @@
-import React from 'react';
+import React, { useState } from 'react';
 import styled from 'styled-components';
-import CreateRoomForm from './CreateRoomForm';
+import axios from 'axios';
+import { useDispatch } from 'react-redux';
+import { addRoom } from '../../redux/modules/roomSlice';
 
 function ModalCreateRoom({ onClose }) {
   const closeModal = () => {
     onClose();
+  };
+
+  const [roomName, setRoomName] = useState('');
+  const [category, setCategory] = useState('초기값^^');
+  const dispatch = useDispatch();
+
+  const handleSelectChange = (e) => {
+    setCategory(e.target.value);
+  };
+
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    try {
+      await axios.post('http://localhost:3001/rooms', {
+        roomName,
+        category,
+      });
+      window.location.reload();
+      //FIXME: dispatch를 써야할지 안써도 될지 고민중..
+      // dispatch(addRoom({ roomName, category }));
+    } catch (error) {
+      console.error(error);
+    }
   };
 
   return (
@@ -16,7 +41,29 @@ function ModalCreateRoom({ onClose }) {
       {/* 상단바.. */}
       <StDivider />
       {/* 방이름,  카테고리, 방만들기,취소하기 버튼. 컴포넌트 */}
-      <CreateRoomForm />
+      {/* <CreateRoomForm /> */}
+      <form onSubmit={handleSubmit}>
+        <div>
+          <h3>방이름</h3>
+          <input
+            type='text'
+            value={roomName}
+            onChange={(e) => setRoomName(e.target.value)}
+          />
+        </div>
+
+        <h3>카테고리</h3>
+        <select value={category} onChange={handleSelectChange}>
+          <option>카테1</option>
+          <option>카테2</option>
+          <option>카테3</option>
+          <option>카테4</option>
+        </select>
+        <div>
+          <button onClick={closeModal}>취소하기</button>
+          <button type='submit'>방만들기</button>
+        </div>
+      </form>
     </StContainer>
   );
 }
